Threshold search multi-threaded, improved GUI
The threshold search is now multi-threaded, and improved in that it
looks for the best circles rather than the first ones it encounters.
The GUI also now allows for adjusting neighbor overlap, and reverse
search has been removed since it is no longer relevant, and allow
overlap, has been replaced with and adjustable neighbor overlap
parameter.