Fixed Build LUT bugs
There were two bugs in the original implementation of the build LUT
functions.

1) The function would exclude anlges of 90, 180, and 270 because rcos or
rsin would take on a value of zero.

2) The function would not reject redundant coordinates.  This is because
it was comparing the coordinates against the next empty slot in the
array, not the previous slot, so it would always accept the new
coordinates, even if they were exactly the same as the previous ones.