Fixed harmonic ghost image and non-divisor problem
The following bugs were fixed:
If the min radius was used as the max resolution, then a high scoreing
ghost image would appear in the next few transforms due to harmoics in
the transform.  By pushing the resolution cutoff to the max radius, this
pushing the ghost image harmonic just outside the transform space.

Another issue was that a user inputting an increment that was not a
divisor of the radius range, it would throw and exception.  Now the
plugin checks for this problem, and instead exits, returning an error.