Fixed bug with single radius transforms
If the transform was performed on a single radius rather than a range,
the plugin would crash when trying to convert the stack to a hyperstack,
since the radius dimension (z) only had one slice.  Therefore, the
plugin only makes a hyperstack if there is more than one slice per
frame.