Enhanced GUI with Easy Mode, fixed multiple GUI use bug
Enhanced GUI with reduced number of options in Easy mode (assumes
highest values for all other values), and only shows local options in
local mode.

Also, fixed bug where plugin would crash if GUI was run twice in a row.
This was because nCircles was not being reset between each run, so now
nCircles is initialized with each run of the plugin.